firmware: configs: moved constant configuration to configs folder
